Separate Reloader tests from other tests to better ensure testing fidelity.
